fix(wdi): save static copy of grapher metadata rather than dynamically fetching it
